[UI v2] feat: Moves Typography component to use cva style
