+++ ui-v2/src/components/ui/typography/index.ts
@@ -1,0 +1,1 @@
+export { Typography } from "./typography";
